+++ kernel/klib/c_helpers/printf_fns.h
@@ -2,8 +2,8 @@
 #define __KLIB_PRINTF_TYPE_FN_H
 
 #ifndef AZALEA_TEST_CODE
-int klib_snprintf(char *out_str, int max_out_len, const char *fmt, ...);
-int klib_vsnprintf(char *out_str, int max_out_len, const char *fmt, va_list args);
+uint32_t klib_snprintf(char *out_str, uint64_t max_out_len, const char *fmt, ...);
+uint32_t klib_vsnprintf(char *out_str, uint64_t max_out_len, const char *fmt, va_list args);
 #endif
 
 #endif
